+++ cacheiql-client/src/errorhandling.ts
@@ -1,0 +1,10 @@
+import { ClientErrorType } from './types';
+
+// client error function//
+export const createClientError = (message: string): ClientErrorType => {
+    return {
+      log: message,
+      status: 400,
+      message: { err: 'Something went wrong in cacheqlIt fetch' },
+    };
+  };
