--- test-client/distTSC/src/client/components/HitMiss.js
@@ -1,7 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const jsx_runtime_1 = require("react/jsx-runtime");
-const HitMiss = ({ time }) => {
-    return ((0, jsx_runtime_1.jsx)("div", { className: 'timebox', children: (0, jsx_runtime_1.jsx)("p", { children: time }) }));
-};
-exports.default = HitMiss;
