--- test-client/distTSC/src/client/types.js
@@ -1,3 +1,0 @@
-"use strict";
-//you need to create a type for props or you will face errors
-Object.defineProperty(exports, "__esModule", { value: true });
